LayoutTests/imported/w3c/web-platform-tests/css/css-text/overflow-wrap/word-wrap-alias-expected.txt


PASS word-wrap should be defined as an alias of overflow-wrap
PASS word-wrap is a legacy name alias for overflow-wrap
